pyproject.toml
[project]
name = "fact"
version = "0.1.0"
description = "Real-time YouTube live stream fact-checker"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=14.0",
    "openai>=1.60.0",
    "yt-dlp>=2024.12.0",
    "pydantic-settings>=2.7.0",
    "chromadb>=0.6.0",
    "aiosqlite>=0.20.0",
    "python-dotenv>=1.0.0",
    "pypdf>=4.0.0",
]

[project.scripts]
fact = "app.main:main"
